pyproject.toml
[project]
name = "Pytorch-Lightning-uv"
description = "pytorch template"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "A.J.Zeller", email = "hello@atticux.me" }]
maintainers = [{ name = "A.J.Zeller", email = "hello@atticux.me" }]
dynamic = ["version"] # https://hatch.pypa.io/latest/config/metadata/#version
dependencies = [
    "accelerate>=1.11.0",
    "bitsandbytes>=0.48.1",
    "evaluate>=0.4.6",
    "kornia>=0.8.0",
    "lightning>=2.5.0.post0",
    "matplotlib>=3.10.0",
    "pandas>=2.3.3",
    "pillow>=11.1.0",
    "requests>=2.32.3",
    "timm>=1.0.14",
    "torch>=2.1,<=2.9.0",
    "torchmetrics>=1.6.1",
    "torchvision>=0.19.0",
    "transformers>=4.45",
    "typer>=0.15.1",
    "wandb[sweeps]>=0.19.5",
]

[project.optional-dependencies]
# optional deps for package installation

[dependency-groups]
dev = [
    "ruff>=0.6.3", # Linting
    "mypy>=1.11.2", # Type checking
    "pre-commit>=3.8.0", # Pre-commit hooks
    "pytest>=8.3.2", # Testing framework
    "pytest-sugar>=1.0.0", # Nicer pytest output
    "types-pyyaml>=6.0.12.20241230",
    "git-cliff>=2.8.0",
]

[[tool.uv.index]]
name     = "pytorch-cu121"
url      = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

## Test
[tool.mypy]
strict  = false
exclude = ["venv", ".venv", "data"] # ignore dataset preprocessing type checking

[tool.pytest.ini_options]
# Set additional command line options for pytest=
# -r= show extra test summary info
# X= show extra info on xfailed tests
# s= don't capture stdout (allow print statements)
# --strict-config= any warnings about configuration are treated as errors
# --strict-markers= treat unregistered markers as errors
addopts        = "-rXs --strict-config --strict-markers"
xfail_strict   = true                                    # Treat tests that are marked as xfail but pass as test failures
filterwarnings = ["error"]                               # Treat all warnings as errors
pythonpath     = "src/"

## Linter and formatter
[tool.ruff]
# cover and extend the default config in https=//docs.astral.sh/ruff/configuration/
extend-exclude = [""]
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

isort = { combine-as-imports = true, split-on-trailing-comma = false }

# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

[tool.ruff.lint.pydocstyle]
convention = "numpy" # Accepts: "google", "numpy", or "pep257".

[tool.ruff.format]
docstring-code-format     = true
skip-magic-trailing-comma = true

# Reference
# 1. https=//github.com/Kludex/python-template/blob/main/template/%7B%7B%20project_slug%20%7D%7D/pyproject.toml.jinja
# 2. https=//github.com/fastapi/full-stack-fastapi-template/blob/master/backend/pyproject.toml
# 3. https=//github.com/pydantic/logfire
# 4. https=//coverage.readthedocs.io/en/latest/index.html

## VCS
[tool.git-cliff.remote.github]
owner = "atticuszeller"
repo  = "Pytorch-Lightning-uv"

[tool.git-cliff.changelog]
# template for the changelog header
header = """
# Changelog\n
All notable changes to this project will be documented in this file.\n
"""
# template for the changelog body
# https://keats.github.io/tera/docs/#introduction
body = """
{% if version %}\
    ## {{ version | trim_start_matches(pat="v") }} - {{ timestamp | date(format="%Y-%m-%d") }}
{% else %}\
    ## unreleased
{% endif %}\
{% for group, commits in commits | group_by(attribute="group") %}
    ### {{ group | striptags | trim | upper_first }}
    {% for commit in commits| unique(attribute="message") %}
        - {% if commit.scope %}*({{ commit.scope }})* {% endif %}\
            {% if commit.breaking %}[**breaking**] {% endif %}\
            {{ commit.message | upper_first }}\
            {% if commit.remote.pr_number %} in #{{ commit.remote.pr_number }}{%- endif %}\
    {% endfor %}
{% endfor %}\n
"""
# template for the changelog footer
footer = """
<!-- generated by git-cliff -->
"""
# remove the leading and trailings
trim = true
# postprocessors
# postprocessors = [
#   { pattern = '<REPO>', replace = "https://github.com/atticuszeller/Pytorch-Lightning-uv" }, # replace repository URL
# ]
# render body even when there are no releases to process
render_always = true
# output file path
output = "CHANGELOG.md"

[tool.git-cliff.git]
# parse the commits based on https://www.conventionalcommits.org
conventional_commits = true
# filter out the commits that are not conventional
filter_unconventional = true
# process each line of a commit as an individual commit
split_commits = false
# regex for preprocessing the commit messages
commit_preprocessors = [
    # If the spelling is incorrect, it will be automatically fixed.
    { pattern = '.*', replace_command = 'typos --write-changes -' },
]
# regex for parsing and grouping commits
commit_parsers = [
    { message = "^feat", group = "<!-- 0 -->🚀 Features" },
    { message = "^fix", group = "<!-- 1 -->🐛 Bug Fixes" },
    { message = "^doc", group = "<!-- 3 -->📚 Documentation" },
    { message = "^perf", group = "<!-- 4 -->⚡ Performance" },
    { message = "^refactor", group = "<!-- 2 -->🚜 Refactor" },
    { message = "^style", group = "<!-- 5 -->🎨 Styling" },
    { message = "^test", group = "<!-- 6 -->🧪 Testing" },
    { message = "^chore\\(release\\)", skip = true },
    { message = "^chore\\(deps.*\\)", skip = true },
    { message = "^chore\\(pr\\)", skip = true },
    { message = "^chore\\(pull\\)", skip = true },
    { message = "^chore|^ci", group = "<!-- 7 -->⚙️ Miscellaneous Tasks" },
    { body = ".*security", group = "<!-- 8 -->🛡️ Security" },
    { message = "^revert", group = "<!-- 9 -->◀️ Revert" },
]
# filter out the commits that are not matched by commit parsers
filter_commits = false
# sort the tags topologically
topo_order = false
# sort the commits inside sections by oldest/newest order
sort_commits = "oldest"

[tool.bumpversion]
current_version        = "0.2.2"
parse                  = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize              = ["{major}.{minor}.{patch}"]
search                 = "{current_version}"
replace                = "{new_version}"
regex                  = false
ignore_missing_version = false
ignore_missing_files   = false
tag                    = true
sign_tags              = false
tag_name               = "v{new_version}"
tag_message            = "chore(release): {current_version} → {new_version}"
allow_dirty            = true                                                  # git-cliff first then bump patch
commit                 = true
message                = "chore(release): {current_version} → {new_version}"
commit_args            = ""
setup_hooks            = []
pre_commit_hooks       = []
post_commit_hooks      = []

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search   = "unreleased"
replace  = "{new_version} - {now:%Y-%m-%d}"

# https://callowayproject.github.io/bump-my-version/reference/search-and-replace-config/
